requirements.txt
# Azure Doc Agent with Skills and MCP
# Python Dependencies

## Core Dependencies
openai>=1.12.0
httpx>=0.26.0
python-dotenv>=1.0.0

## MCP Client - Official SDK
mcp>=1.0.0

## Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

## Azure Dependencies
azure-identity>=1.15.0
agent-framework>=0.1.0

## Async Support
asyncio>=3.4.3
aiohttp>=3.9.0

## Logging
colorlog>=6.8.0

## Development Dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
